Make menus in window captures visible.
Unfortunately menus do not seem to belong to windows itself, so capture
the window by capturing a cropped version of the root window. Most menus
should be visible like that.
Only do so while the window is active though, otherwise we might capture
stuff that is not supposed to be captured.