Adjust HTML styling to separate UI components
To help the video stand out from the page background, a slight texture
is added to the page background.  The settings UI is adjusted to stand
above the rest.  In light mode, a drop shadow appears below it.  In
dark mode, a slight transparency lets you see the video below it.