Limit the length of the sourceBuffer
By default, the sourceBuffer can grow to several hundred MB, depending
on browser settings.  This can decrease performance and (apparently)
occasionally crash browser tabs.  However, we only care about the end
of the buffer, so most of this use is completely wasted on us.  Ideally,
there would be a setting that said, don't keep frames after they've
been displayed, but if there is, I haven't found it.

This solution sets a target buffer length, 10s as a more-or-less
random value.  (Browsers often got 100s of buffer, so this is 1/10 of
that.  Realistically, anything over a second is probably useless from
a UX experience, so 10s sits in the middle.  We also need to make sure
this is longer than the time between keyframes.)  When the buffer
length exceeds twice the target length, it is trimmed to the target
length.  In limited testing, this seems to improve memory performance,
and I have not seen a crashed tab.