No new buffer doesn't always imply an error.
PipeWire only provides new buffers if the source has been damaged, that
is if the compositor decided to update it. So if there is no new buffer
but still a reference to an old buffer that just means the source has
not changed.

In order to still deliver an acceptable framerate (some clients like
firefox on android do not like videos with a very low framerate) wait
33ms (~30 frames per second) and then just send the old buffer again.